+++ packages/txwrapper-core/src/core/metadata/getRegistryBase.ts
@@ -1,9 +1,13 @@
 import { TypeRegistry } from '@polkadot/types';
 import { ExtDef } from '@polkadot/types/extrinsic/signedExtensions/types';
-import { AnyJson, RegistryTypes } from '@polkadot/types/types';
+import {
+	AnyJson,
+	OverrideBundleType,
+	RegistryTypes,
+} from '@polkadot/types/types';
 
 import { ChainProperties } from '../../types';
-import { createMetadata } from './createMetadata';
+import { createMetadataUnmemoized } from './createMetadata';
 
 export interface GetRegistryBaseArgs {
 	/**
@@ -30,6 +34,14 @@ export interface GetRegistryBaseArgs {
 	 * User extensions used to inject into the type registry
 	 */
 	userExtensions?: ExtDef;
+	/**
+	 * OverrideTypesBundle to set to registry
+	 */
+	typesBundle?: OverrideBundleType;
+	/**
+	 * Additional types to register in the registry.
+	 */
+	additionalTypes?: RegistryTypes;
 }
 
 /**
@@ -42,10 +54,11 @@ export function getRegistryBase({
 	asCallsOnlyArg,
 	signedExtensions,
 	userExtensions,
+	typesBundle,
+	additionalTypes,
 }: GetRegistryBaseArgs): TypeRegistry {
 	const registry = new TypeRegistry();
-
-	const generatedMetadata = createMetadata(
+	const generatedMetadata = createMetadataUnmemoized(
 		registry,
 		metadataRpc,
 		asCallsOnlyArg
@@ -53,7 +66,13 @@ export function getRegistryBase({
 
 	registry.register(specTypes);
 
+	if (additionalTypes) {
+		registry.register(additionalTypes);
+	}
+
 	registry.setMetadata(generatedMetadata, signedExtensions, userExtensions);
+
+	registry.setKnownTypes({ typesBundle });
 
 	// Register the chain properties for this registry
 	registry.setChainProperties(