+++ packages/txwrapper-dev/src/metadata/polkadot/index.ts
@@ -1,1 +1,0 @@
-export * from './polkadotV9122MetadataHex';