+++ packages/txwrapper-dev/src/metadata/static/index.ts
@@ -1,2 +1,0 @@
-export * from './staticV3-1-1';
-export * from './staticV4-3-1';