+++ packages/txwrapper-substrate/src/methods/proxy/removeProxies.spec.ts
@@ -1,6 +1,6 @@
 import {
 	itHasCorrectBaseTxInfo,
-	POLKADOT_25_TEST_OPTIONS,
+	KUSAMA_TEST_OPTIONS,
 	TEST_BASE_TX_INFO,
 } from '@substrate/txwrapper-dev';
 
@@ -8,14 +8,10 @@ import { removeProxies } from './removeProxies';
 
 describe('proxy::removeProxies', () => {
 	it('should work', () => {
-		const unsigned = removeProxies(
-			{},
-			TEST_BASE_TX_INFO,
-			POLKADOT_25_TEST_OPTIONS
-		);
+		const unsigned = removeProxies({}, TEST_BASE_TX_INFO, KUSAMA_TEST_OPTIONS);
 
 		itHasCorrectBaseTxInfo(unsigned);
 
-		expect(unsigned.method).toBe('0x1d03');
+		expect(unsigned.method).toBe('0x1e03');
 	});
 });