+++ packages/txwrapper-substrate/src/methods/system/remark.spec.ts
@@ -1,6 +1,6 @@
 import {
 	itHasCorrectBaseTxInfo,
-	POLKADOT_25_TEST_OPTIONS,
+	KUSAMA_TEST_OPTIONS,
 	TEST_BASE_TX_INFO,
 } from '@substrate/txwrapper-dev';
 
@@ -12,10 +12,10 @@ describe('system::remark', () => {
 		const unsigned = remark(
 			TEST_METHOD_ARGS.system.remark,
 			TEST_BASE_TX_INFO,
-			POLKADOT_25_TEST_OPTIONS
+			KUSAMA_TEST_OPTIONS
 		);
 
 		itHasCorrectBaseTxInfo(unsigned);
-		expect(unsigned.method).toBe('0x00011042194253');
+		expect(unsigned.method).toBe('0x00001042194253');
 	});
 });