+++ packages/txwrapper-substrate/src/methods/vesting/vest.spec.ts
@@ -1,6 +1,6 @@
 import {
 	itHasCorrectBaseTxInfo,
-	POLKADOT_25_TEST_OPTIONS,
+	KUSAMA_TEST_OPTIONS,
 	TEST_BASE_TX_INFO,
 } from '@substrate/txwrapper-dev';
 
@@ -12,10 +12,10 @@ describe('vesting::vest', () => {
 		const unsigned = vest(
 			TEST_METHOD_ARGS.vesting.vest,
 			TEST_BASE_TX_INFO,
-			POLKADOT_25_TEST_OPTIONS
+			KUSAMA_TEST_OPTIONS
 		);
 
 		itHasCorrectBaseTxInfo(unsigned);
-		expect(unsigned.method).toBe('0x1b00');
+		expect(unsigned.method).toBe('0x1c00');
 	});
 });