+++ packages/txwrapper-substrate/src/test-helpers/TEST_METHOD_ARGS.ts
@@ -31,11 +31,11 @@ export const TEST_METHOD_ARGS = {
 	},
 	balances: {
 		transferKeepAlive: {
-			dest: 'Fy2rsYCoowQBtuFXqLE65ehAY9T6KWcGiNCQAyPDCkfpm4s',
+			dest: { id: 'Fy2rsYCoowQBtuFXqLE65ehAY9T6KWcGiNCQAyPDCkfpm4s' },
 			value: 12,
 		},
 		transferAll: {
-			dest: 'Fy2rsYCoowQBtuFXqLE65ehAY9T6KWcGiNCQAyPDCkfpm4s',
+			dest: { id: 'Fy2rsYCoowQBtuFXqLE65ehAY9T6KWcGiNCQAyPDCkfpm4s' },
 			keepAlive: true,
 		},
 	},
@@ -70,19 +70,19 @@ export const TEST_METHOD_ARGS = {
 			delegate: 'HNZata7iMYWmk5RvZRTiAsSDhV8366zq2YGb3tLH5Upf74F', // seed "//Alice"
 			real: '14E5nqKAp3oAJcmzgZhUD2RcptBeUBScxKHgJKU4HPNcKVf3', // seed "//Bob",
 			forceProxyType: 'Any',
-			call: '0x0500306721211d5404bd9da88e0204360a1a9ab8b87c66c1bc2fcdd37f3c2222cc200f00a0be1c448399',
+			call: '0x0400004aca27604ad033f7c45b1cfc23b55520826db4abb69a8a7c165461c40f330c6b0700e8764817',
 		},
 		proxy: {
 			real: '14E5nqKAp3oAJcmzgZhUD2RcptBeUBScxKHgJKU4HPNcKVf3', // seed "//Bob",
 			forceProxyType: 'Any',
-			call: '0x0500306721211d5404bd9da88e0204360a1a9ab8b87c66c1bc2fcdd37f3c2222cc200f00a0be1c448399',
+			call: '0x0400004aca27604ad033f7c45b1cfc23b55520826db4abb69a8a7c165461c40f330c6b0700e8764817',
 		},
-		anonymous: {
+		createPure: {
 			proxyType: 'Any',
 			delay: 30,
 			index: 1,
 		},
-		killAnonymous: {
+		killPure: {
 			spawner: '14E5nqKAp3oAJcmzgZhUD2RcptBeUBScxKHgJKU4HPNcKVf3',
 			proxyType: 'Any',
 			index: 0,
@@ -92,13 +92,19 @@ export const TEST_METHOD_ARGS = {
 	},
 	session: {
 		setKeys: {
-			keys: [
-				'HNZata7iMYWmk5RvZRTiAsSDhV8366zq2YGb3tLH5Upf74F', // seed "//Alice"
-				'FoQJpPyadYccjavVdTWxpxU7rUEaYhfLCPwXgkfD6Zat9QP', // seed "//Bob"
-				'Fr4NzY1udSFFLzb2R3qxVQkwz9cZraWkyfH4h3mVVk7BK7P', // seed "//Charlie"
-				'HNZata7iMYWmk5RvZRTiAsSDhV8366zq2YGb3tLH5Upf74F', // seed "//Alice"
-				'FoQJpPyadYccjavVdTWxpxU7rUEaYhfLCPwXgkfD6Zat9QP', // seed "//Bob"
-			],
+			keys: {
+				authority_discovery:
+					'0x6c28300123ec5d172f707da0e387c066b47cc38198a41beeaf476ca24ac4f46a',
+				babe: '0x0c635d996fb571f17b26a7ed139a52c71ef7161f07932d21bddeec8426647023',
+				grandpa:
+					'0xbe0bccde60dcf2dee2562bc249b5c87a892b551b7058c2a4761ad05f57517ab7',
+				im_online:
+					'0xbc29159bd6a6686e825bb16043903c461b61f682187d5ea50c59241526d96b43',
+				para_assignment:
+					'0xa60957851edecb59585aec372d109b7ee5461cc7da38e70b24bf20cd126fa83f',
+				para_validator:
+					'0xe2459cc64f42474adefc2108f938c594965127f2685002dcca69e25d4f5a8636',
+			},
 			proof: '0x',
 		},
 	},
@@ -190,7 +196,6 @@ export const TEST_METHOD_ARGS = {
 				index: 3,
 			},
 			call: '0x040300cc5fc7439d3c49984e7345319de148561cd930748fb401276d289f714acdd50c28',
-			storeCall: false,
 			maxWeight: { refTime: '90071992547409910', proofSize: '0' },
 		},
 	},