+++ packages/txwrapper-polkadot/CHANGELOG.md
@@ -3,6 +3,67 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0](https://github.com/paritytech/txwrapper-core/compare/v5.0.1...v6.0.0) (2023-05-08)
+
+
+* fix!: refactor the whole testing system with metadata, remove old calls, and fix some types (#295) ([58d026a](https://github.com/paritytech/txwrapper-core/commit/58d026ad7c0d9eaa0816fddf33735d4015e22edd)), closes [#295](https://github.com/paritytech/txwrapper-core/issues/295)
+
+
+### Features
+
+* add `additionalTypes`, and `typesBundle` to `getRegistryBase` ([#294](https://github.com/paritytech/txwrapper-core/issues/294)) ([671a2d1](https://github.com/paritytech/txwrapper-core/commit/671a2d168a114866455aff2f01d0d019eeeed2d3))
+
+
+### BREAKING CHANGES
+
+* fix balances
+
+* metadata and decode
+
+* method
+
+* remove constants
+
+* remove a large amount of bloat
+
+* ensure setKets kets type is correct
+
+
+
+
+
+## [5.0.1](https://github.com/paritytech/txwrapper-core/compare/v5.0.0...v5.0.1) (2023-02-28)
+
+**Note:** Version bump only for package @substrate/txwrapper-polkadot
+
+
+
+
+
+# [5.0.0](https://github.com/paritytech/txwrapper-core/compare/v4.1.0...v5.0.0) (2023-01-13)
+
+**Note:** Version bump only for package @substrate/txwrapper-polkadot
+
+
+
+
+
+# [4.1.0](https://github.com/paritytech/txwrapper-core/compare/v4.0.3...v4.1.0) (2022-12-29)
+
+**Note:** Version bump only for package @substrate/txwrapper-polkadot
+
+
+
+
+
+## [4.0.3](https://github.com/paritytech/txwrapper-core/compare/v4.0.2...v4.0.3) (2022-12-07)
+
+**Note:** Version bump only for package @substrate/txwrapper-polkadot
+
+
+
+
+
 ## [4.0.2](https://github.com/paritytech/txwrapper-core/compare/v4.0.1...v4.0.2) (2022-11-28)
 
 **Note:** Version bump only for package @substrate/txwrapper-polkadot