+++ packages/txwrapper-substrate/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0](https://github.com/paritytech/txwrapper-core/compare/v5.0.1...v6.0.0) (2023-05-08)
+
+
+* fix!: refactor the whole testing system with metadata, remove old calls, and fix some types (#295) ([58d026a](https://github.com/paritytech/txwrapper-core/commit/58d026ad7c0d9eaa0816fddf33735d4015e22edd)), closes [#295](https://github.com/paritytech/txwrapper-core/issues/295)
+
+
+### BREAKING CHANGES
+
+* fix balances
+
+* metadata and decode
+
+* method
+
+* remove constants
+
+* remove a large amount of bloat
+
+* ensure setKets kets type is correct
+
+
+
+
+
+## [5.0.1](https://github.com/paritytech/txwrapper-core/compare/v5.0.0...v5.0.1) (2023-02-28)
+
+**Note:** Version bump only for package @substrate/txwrapper-substrate
+
+
+
+
+
+# [5.0.0](https://github.com/paritytech/txwrapper-core/compare/v4.1.0...v5.0.0) (2023-01-13)
+
+
+### Bug Fixes
+
+* remove `store_call` from `asMulti` ([#280](https://github.com/paritytech/txwrapper-core/issues/280)) ([c4ce505](https://github.com/paritytech/txwrapper-core/commit/c4ce505fa96917ac20da5aa5621196fbff83fd38))
+
+
+
+
+
+# [4.1.0](https://github.com/paritytech/txwrapper-core/compare/v4.0.3...v4.1.0) (2022-12-29)
+
+
+### Features
+
+* Add create pure and kill pure transaction for version > 0.9.30 ([#276](https://github.com/paritytech/txwrapper-core/issues/276)) ([8af876e](https://github.com/paritytech/txwrapper-core/commit/8af876ee76e2c572f16c06ca838eecd020eaa3b9))
+
+
+
+
+
+## [4.0.3](https://github.com/paritytech/txwrapper-core/compare/v4.0.2...v4.0.3) (2022-12-07)
+
+
+### Bug Fixes
+
+* **bug:** give WeightsV2 support for multisig  ([#269](https://github.com/paritytech/txwrapper-core/issues/269)) ([82150bb](https://github.com/paritytech/txwrapper-core/commit/82150bb06ccad6ef691affab98922266efe0e48b))
+
+
+
+
+
 ## [4.0.2](https://github.com/paritytech/txwrapper-core/compare/v4.0.1...v4.0.2) (2022-11-28)
 
 **Note:** Version bump only for package @substrate/txwrapper-substrate